+++ btcclient/zmq/subscribe.go
@@ -177,5 +177,5 @@ func (c *Client) getBlockByHash(blockHash *chainhash.Hash) (*types.IndexedBlock,
 	}
 
 	btcTxs := types.GetWrappedTxs(mBlock)
-	return types.NewIndexedBlock(int32(blockInfo.Height), &mBlock.Header, btcTxs), mBlock, nil
+	return types.NewIndexedBlock(blockInfo.Height, &mBlock.Header, btcTxs), mBlock, nil
 }
